chore: move e2e endpoints
